+++ src/test/java/com/knubisoft/testlum/testing/framework/db/s3/S3Operation.java
@@ -8,6 +8,7 @@ import com.knubisoft.testlum.testing.framework.db.StorageOperation;
 import com.knubisoft.testlum.testing.framework.db.source.Source;
 import com.knubisoft.testlum.testing.framework.env.AliasEnv;
 import com.knubisoft.testlum.testing.framework.env.EnvManagerImpl.EnvProvider;
+import com.knubisoft.testlum.testing.model.global_config.S3;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.stereotype.Component;
@@ -17,7 +18,7 @@ import java.util.Objects;
 
 @Conditional({OnS3EnabledCondition.class})
 @Component
-public class S3Operation implements StorageOperation {
+public class S3Operation extends StorageOperation {
 
     private final Map<AliasEnv, AmazonS3> amazonS3;
 
@@ -33,10 +34,11 @@ public class S3Operation implements StorageOperation {
     @Override
     public void clearSystem() {
         this.amazonS3.forEach((aliasEnv, amazonS3) -> {
-            if (Objects.equals(aliasEnv.getEnvironment(), EnvProvider.currentEnv())) {
+            if (isTruncate(S3.class, aliasEnv) && Objects.equals(aliasEnv.getEnvironment(), EnvProvider.currentEnv())) {
                 amazonS3.listBuckets().forEach(bucket -> {
                     ListObjectsV2Result objectsInBucket = amazonS3.listObjectsV2(bucket.getName());
                     this.deleteObjectsInBucket(amazonS3, objectsInBucket, bucket.getName());
+                    amazonS3.deleteBucket(bucket.getName());
                 });
             }
         });
